+++ miner/src/Arktika.java
@@ -46,7 +46,7 @@ import io.grpc.ServerBuilder;
 import snowblossom.client.WalletUtil;
 
 
-public class Arktika
+public class Arktika implements PoolClientOperator
 {
   private static final Logger logger = Logger.getLogger("snowblossom.miner");
 
@@ -73,9 +73,6 @@ public class Arktika
   }
 
   private volatile WorkUnit last_work_unit;
-
-  private MiningPoolServiceStub asyncStub;
-  protected MiningPoolServiceBlockingStub blockingStub;
 
   private final NetworkParams params;
 
@@ -105,6 +102,7 @@ public class Arktika
   private final int layer_count;
   private Stubo stubo;
 
+  protected PoolClientFace pool_client;
 
   public Arktika(Config config) throws Exception
   {
@@ -121,14 +119,7 @@ public class Arktika
 
     params = NetworkParams.loadFromConfig(config);
 
-    if ((!config.isSet("mine_to_address")) && (!config.isSet("mine_to_wallet")))
-    {
-      throw new RuntimeException("Config must either specify mine_to_address or mine_to_wallet");
-    }
-    if ((config.isSet("mine_to_address")) && (config.isSet("mine_to_wallet")))
-    {
-      throw new RuntimeException("Config must either specify mine_to_address or mine_to_wallet, not both");
-    }
+    pool_client = new PoolClient(config, this);
 
     // this is a bad idea, don't use this.  It eats all the cpu doing
     // record keeping
@@ -140,7 +131,7 @@ public class Arktika
 
     loadField();
 
-    subscribe();
+    pool_client.subscribe();
 
     stubo = new Stubo(composit_source, selected_field);
 
@@ -161,71 +152,10 @@ public class Arktika
 
   private ManagedChannel channel;
 
-  private void subscribe() throws Exception
-  {
-    if (channel != null)
-    {
-      channel.shutdownNow();
-      channel = null;
-    }
-
-    String host = config.get("pool_host");
-    int port = config.getIntWithDefault("pool_port", 23380);
-    channel = ManagedChannelBuilder.forAddress(host, port).usePlaintext(true).build();
-
-    asyncStub = MiningPoolServiceGrpc.newStub(channel);
-    blockingStub = MiningPoolServiceGrpc.newBlockingStub(channel);
-
-    AddressSpecHash to_addr = getMineToAddress();
-    String address_str = AddressUtil.getAddressString(params.getAddressPrefix(), to_addr);
-
-    String client_id = null;
-    if (config.isSet("mining_client_id"))
-    {
-      client_id = config.get("mining_client_id");
-    }
-    GetWorkRequest.Builder req = GetWorkRequest.newBuilder();
-    if (client_id != null) req.setClientId(client_id);
-
-    req.setPayToAddress(address_str);
-
-    asyncStub.getWork( req.build(), new WorkUnitEater());
-    logger.info("Subscribed to work");
-
-  }
-
-  private AddressSpecHash getMineToAddress() throws Exception
-  {
-
-    if (config.isSet("mine_to_address"))
-    {
-      String address = config.get("mine_to_address");
-      AddressSpecHash to_addr = new AddressSpecHash(address, params);
-      return to_addr;
-    }
-    if (config.isSet("mine_to_wallet"))
-    {
-      File wallet_path = new File(config.get("mine_to_wallet"));
-      WalletDatabase wallet = WalletUtil.loadWallet(wallet_path, false, params);
-
-      if (wallet.getAddressesCount() == 0)
-      {
-        throw new RuntimeException("Wallet has no addresses");
-      }
-      LinkedList<AddressSpec> specs = new LinkedList<AddressSpec>();
-      specs.addAll(wallet.getAddressesList());
-      Collections.shuffle(specs);
-
-      AddressSpec spec = specs.get(0);
-      AddressSpecHash to_addr = AddressUtil.getHashForSpec(spec);
-      return to_addr;
-    }
-    return null;
-  }
-
   public void stop()
   {
     terminate = true;
+    pool_client.stop();
   }
   public boolean isTerminated()
   {
@@ -273,7 +203,7 @@ public class Arktika
         logger.info("Stalled.  No valid work unit, reconnecting to pool");
         try
         {
-          subscribe();
+          pool_client.subscribe();
         }
         catch (Throwable t)
         {
@@ -335,15 +265,10 @@ public class Arktika
     return wu;
   }
 
-  public class WorkUnitEater implements StreamObserver<WorkUnit>
+  @Override
+  public void notifyNewWorkUnit(WorkUnit wu)
   {
-    public void onCompleted() {}
-    public void onError(Throwable t) 
-    {
-      logger.info("Error talking to mining pool: " + t);
-    }
-    public void onNext(WorkUnit wu)
-    {
+
       int last_block = -1;
       WorkUnit wu_old = last_work_unit;
       if (wu_old != null)
@@ -373,23 +298,23 @@ public class Arktika
 
         last_work_unit = wu_new;
 
-        // If the block number changes, clear the queues
-        if (last_block != wu_new.getHeader().getBlockHeight())
-        {
-          for(FaQueue q : layer_to_queue_map.values())
-          {
-            q.clear();
-          }
-        }
-        
       }
       catch (Throwable t)
       {
         logger.info("Work block load error: " + t.toString());
         last_work_unit = null;
       }
-    }
   }
+
+  @Override
+  public void notifyNewBlock(int height)
+  {
+        for(FaQueue q : layer_to_queue_map.values())
+        {
+          q.clear();
+        }
+  }
+
 
 
   private void loadField() throws Exception